requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
decorator==4.0.11
Django==1.11
django-braces==1.11.0
django-celery==3.2.1
django-crispy-forms==1.6.1
django-messages-to-bootstrap-notify==0.2.3
django-pagination==1.0.7
django-photologue==3.7
django-popup-view-field==0.3.0
django-sortedm2m==1.3.3
django-tabbed-admin==1.0.1
django-tables2==1.6.1
django-widget-tweaks==1.4.1
djangorestframework==3.6.2
epoch==0.1.4
ExifRead==2.1.2
googlemaps==2.4.6
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.3.0
kombu==3.0.37
nbformat==4.3.0
nodejs==0.1.1
olefile==0.44
optional-django==0.1.0
pillow==4.1.1
plotly==2.0.8
psycopg2==2.7.1
pubnub==4.0.10
pycryptodomex==3.4.5
pystache==0.5.4
pytz==2017.2
redis==2.10.5
requests==2.11.1
selenium==3.4.1
six==1.10.0
South==1.0
traitlets==4.3.2
vine==1.1.3
virtualenv==15.1.0
